refactored ecompiler
